stimulus_aiu_backend/requirements.txt
asgiref==3.11.0
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docx==0.2.4
docxtpl==0.20.2
drf-yasg==1.21.11
et_xmlfile==2.0.0
gunicorn==23.0.0
inflection==0.5.1
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
openpyxl==3.1.5
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
